[refactor] coderhub_eda.py renamed to coderhub_stats.py, more comments added
